refactor(gamelift-queue): convert to libnuke resource format
